belajar middleware
membuat auth middleware beserta dengan form login

+++ routes/web.php
@@ -15,35 +15,42 @@ Route::get('/', function () {
     return view('welcome');
 });
 
+
+Route::get('/login', 'Auth\LoginController@form')->name("login");
+Route::post('/login', 'Auth\LoginController@login');
+Route::get('/logout', 'Auth\LoginController@logout')->name("logout");
+
 // ==================== Product ===================== //
 
-Route::get('/product', 'ProductController@index');
-Route::get('/product/form', 'ProductController@create');
-Route::get('/product/form/{product}', 'ProductController@edit');
-Route::get('/product/{product}', 'ProductController@show');
+Route::get('/home', 'ProductController@index')->middleware('auth');
 
-Route::post('/product', 'ProductController@store');
-Route::put('/product/{product}', 'ProductController@update');
-Route::delete('/product/{product}', 'ProductController@destroy');
+Route::get('/product', 'ProductController@index')->middleware('auth');
+Route::get('/product/form', 'ProductController@create')->middleware('auth');
+Route::get('/product/form/{product}', 'ProductController@edit')->middleware('auth');
+Route::get('/product/{product}', 'ProductController@show')->middleware('auth');
+
+Route::post('/product', 'ProductController@store')->middleware('auth');
+Route::put('/product/{product}', 'ProductController@update')->middleware('auth');
+Route::delete('/product/{product}', 'ProductController@destroy')->middleware('auth');
 
 // ==================== Category ===================== //
 
-Route::get('/category', 'CategoryController@index');
-Route::get('/category/form', 'CategoryController@create');
-Route::get('/category/form/{category}', 'CategoryController@edit');
-Route::get('/category/{category}', 'CategoryController@show');
+Route::get('/category', 'CategoryController@index')->middleware('auth');
+Route::get('/category/form', 'CategoryController@create')->middleware('auth');
+Route::get('/category/form/{category}', 'CategoryController@edit')->middleware('auth');
+Route::get('/category/{category}', 'CategoryController@show')->middleware('auth');
 
-Route::post('/category', 'CategoryController@store');
-Route::put('/category/{category}', 'CategoryController@update');
-Route::delete('/category/{category}', 'CategoryController@destroy');
+Route::post('/category', 'CategoryController@store')->middleware('auth');
+Route::put('/category/{category}', 'CategoryController@update')->middleware('auth');
+Route::delete('/category/{category}', 'CategoryController@destroy')->middleware('auth');
 
 // ==================== User ===================== //
 
-Route::get('/user', 'UserController@index');
-Route::get('/user/form', 'UserController@create');
-Route::get('/user/form/{user}', 'UserController@edit');
-Route::get('/user/{user}', 'UserController@show');
+Route::get('/user', 'UserController@index')->middleware('auth');
+Route::get('/user/form', 'UserController@create')->middleware('auth');
+Route::get('/user/form/{user}', 'UserController@edit')->middleware('auth');
+Route::get('/user/{user}', 'UserController@show')->middleware('auth');
 
-Route::post('/user', 'UserController@store');
-Route::put('/user/{user}', 'UserController@update');
-Route::delete('/user/{user}', 'UserController@destroy');
+Route::post('/user', 'UserController@store')->middleware('auth');
+Route::put('/user/{user}', 'UserController@update')->middleware('auth');
+Route::delete('/user/{user}', 'UserController@destroy')->middleware('auth');
